Update atomic types test cases
